+++ drivers/media/huawei/camera/hwcam_cfgpipeline.c
@@ -1,4 +1,26 @@
-
+/*
+ *  Hisilicon K3 SOC camera driver source file
+ *
+ *  Copyright (C) Huawei Technology Co., Ltd.
+ *
+ * Author:
+ * Email:
+ * Date:	  2013-11-21
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ */
 
 
 #include <linux/anon_inodes.h>
@@ -1007,7 +1029,7 @@ hwcam_cfgpipeline_vo_ioctl32(
 			rc = compat_get_v4l2_event_data(kp, up_p);
 			if (0 != rc)
 				return rc;
-			rc = hwcam_cfgpipeline_vo_ioctl(filep, cmd, (unsigned long)(kp));
+			rc = hwcam_cfgpipeline_vo_ioctl(filep, cmd, (unsigned long)(uintptr_t)(kp));
 			if (0 != rc)
 				return rc;
 			rc = compat_put_v4l2_event_data(kp, up_p);
