+++ include/linux/hisi/hisi_powerkey_event.h
@@ -16,11 +16,11 @@ typedef  enum {
 #ifdef  CONFIG_HISI_POWERKEY_SPMI
 int hisi_powerkey_register_notifier(struct notifier_block *nb);
 int hisi_powerkey_unregister_notifier(struct notifier_block *nb);
-int hisi_call_powerkey_notifiers(int val,void *v);
+int hisi_call_powerkey_notifiers(unsigned long val,void *v);
 #else
 static inline int hisi_powerkey_register_notifier(struct notifier_block *nb){return 0;};
 static inline int hisi_powerkey_unregister_notifier(struct notifier_block *nb){return 0;};
-static inline int hisi_call_powerkey_notifiers(int val,void *v){return 0;};
+static inline int hisi_call_powerkey_notifiers(unsigned long val, void *v){return 0;};
 #endif
 
 #endif
